==> Curso engenharia de dados/Scrip/Python - SGBDS/07. Google BigQuery/1.Conexão com Autentificador Físico.ipynb <==
#pip install  google-cloud-bigquery
#pip install --upgrade google-cloud-bigquery
#pip install  pandas_gbq
#pip install  pandas

from google.cloud import bigquery
import pandas as pd

# Especifique o caminho para o arquivo JSON de credenciais
credencial = r"C:\Users\edmil\Downloads\pythoncurso-398511-260383f13f39.json"

# Crie um cliente BigQuery usando as credenciais
client = bigquery.Client.from_service_account_json(credencial)

# Construa sua consulta
query = """ 

  SELECT 
  id
  ,first_name
  ,last_name
  ,email  
 FROM `pythoncurso-398511.AulaPython.Clientes` 
 where state in ('Piauí','Alagoas','Amazonas')

"""
# Execute a consulta
resultado = client.query(query)

"""
for row in resultado:
    print(row)
"""

# Converta os resultados em um DataFrame
df = resultado.to_dataframe()



df.head(5)
