==> Curso engenharia de dados/Scrip/Python - SGBDS/07. Google BigQuery/3.Conexão com Auyth2 Variavel.ipynb <==
from google.cloud import bigquery
import pandas as pd
from google.oauth2 import service_account

# Especifique as informaçoes de credenciais
credencial = {
  "type": "",
  "project_id": "-",
  "private_key_id": "",
  "private_key": "- PRIVATE KEY-",
  "client_email": "",
  "client_id": "",
  "auth_uri": "https://..com///auth",
  "token_uri": "https://..com/token",
  "auth_provider_x509_cert_url": "..///",
  "client_x509_cert_url": "",
  "universe_domain": ".com"
}

# Crie o cliente BigQuery
client = bigquery.Client(credentials=service_account.Credentials.from_service_account_info(credencial))

# Construa sua consulta
query = """ 

  SELECT 
  id
  ,first_name
  ,last_name
  ,email  
  ,state
 FROM `pythoncurso-398511.AulaPython.Clientes` 
 where state in ('Piauí','Alagoas',
 'Amazonas','Maranhão','Rondônia')
"""
# Execute a consulta
resultado = client.query(query)

# Converta os resultados em um DataFrame
df = resultado.to_dataframe()
df.head(30)
